+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Version History
 
+## 7.2.0
 * [THORN-195] Add HardwareVerificationError.
+* [THORN-214] Added HardwareVersion property to MccsTile. If not defined ADC0 -> ADC15 temperature attributes are not evaluated in health
+* [THORN-214] Added BiosVersion property to MccsTile. If not defined pll_40g attribute is not evaluated in health
+* [THORN-214] Added PreAduPresent property to MccsTile (default == True). When false we expect currents FE0_mVA and FE1_mVA to be 0 with some tolerance.
 
 ## 7.1.0